requirements.txt
# Core Application Dependencies
fastapi[all]
openai
python-dotenv

# Testing & Quality Dependencies
pytest
pytest-mock
ruff